Test environment setup + Game.js addition
I added a seperate client id which allows me to run the program on node server for testing.

Game.js was created and I am beggining work on it. Lobby is all but finished.
I think I want to store the users actual name in storage somewhere for the game screen