Routing working. Basic restructure complete
The basics of how I wanted to restructure the files is done. Now I want to refactor the code and at this point I believe the architecture I will implement is MVVM. I want to minimize the use of firebase services at this point. The current implementation is rife with problems.

Current plan is to use event listners to ensure that the model and view stay in sync which should enable gamepaly.

I may implement a few server side features to clear out old game nodes etc.

I also want to add a navbar at the top of the page, account page, and some user specific tracking.
Potentially will try and switch the login to a modal rather than a redirect.